Add basic LoRaWAN dissector (WIP, cf full message)
The file in packet-iec104.c has been changed in order to take into account
packets intercepted between a gateway and the ThingPark server. Unfortunately, I
then discovered it is actually not LoRaWAN that is transmitted encapsulated in
IEC 60870-5-104 frames, but a proprietary ThingPark protocol. So the diff to
epan/dissectors/packet-iec104.c is actually useless, but it is left there as an
indicator as to how to add a dissector table to a protocol without it, and use
it.

Work still to be done:
 * Test (so, need to find a way to dump actual LoRa frames... [1]?)
 * Handle FOpts
 * Find out where to actually hook the dissector for proper dissection (given
   it's not in packet-iec104.c)
 * Cleanup (remove CMakeListsCustom.txt, unused packet-iec104.c, etc.)
 * Move to a single-file dissector

[1] https://github.com/mirakonta/lora_gateway/wiki